# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_HierarchySummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HierarchySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[HierarchySummary]",
        "type": "object"
      },
...
      "Page_Union_WorkSearchHit__HierarchySummary__": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkSearchHit"
                },
                {
                  "$ref": "#/components/schemas/HierarchySummary"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Union[WorkSearchHit, HierarchySummary]]",
        "type": "object"
      },
...
      "WorkCreation": {
        "additionalProperties": false,
        "properties": {
          "initial_checkpoint": {
            "$ref": "#/components/schemas/CheckpointRead"
          },
          "initial_relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipEdgeRead"
            },
            "title": "Initial Relationships",
            "type": "array"
          },
          "work_item": {
            "$ref": "#/components/schemas/WorkItemRead"
          }
        },
        "required": [
          "work_item",
          "initial_checkpoint"
        ],
        "title": "WorkCreation",
        "type": "object"
      },
...
      "WorkItemCreate": {
        "additionalProperties": false,
        "properties": {
          "client_operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional caller-generated UUID for durable replay of this mutation. Reuse it only with the exact same operation and semantic arguments after an unknown outcome.",
            "title": "Client Operation Id"
          },
          "initial_checkpoint": {
            "$ref": "#/components/schemas/InitialCheckpointCreate"
          },
          "initial_relationships": {
            "items": {
              "$ref": "#/components/schemas/InitialRelationshipCreate"
            },
            "maxItems": 10,
            "title": "Initial Relationships",
            "type": "array"
          },
          "priority": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "status": {
            "default": "pending",
            "enum": [
              "pending",
              "wont-do",
              "promoted"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "summary",
          "initial_checkpoint"
        ],
        "title": "WorkItemCreate",
        "type": "object"
      },
      "WorkItemDetailRead": {
        "additionalProperties": false,
        "properties": {
          "canonical": {
            "$ref": "#/components/schemas/CanonicalWorkProjection"
          },
          "work_item": {
            "$ref": "#/components/schemas/WorkItemRead"
          }
        },
        "required": [
          "work_item",
          "canonical"
        ],
        "title": "WorkItemDetailRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Durable project-scoped work with immutable agent checkpoints.",
    "title": "Mnemonic API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/work-items": {
      "get": {
        "operationId": "search_work_api_v1_projects__project_id__work_items_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "semantic",
            "required": false,
            "schema": {
              "default": false,
              "title": "Semantic",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "active",
                "dropped",
                "deferred",
                "done",
                "wont-do",
                "promoted",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated",
              "enum": [
                "updated",
                "created",
                "priority"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "source_client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Client"
            }
          },
          {
            "in": "query",
            "name": "source_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Session Id"
            }
          },
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "full",
              "enum": [
                "full",
                "roots"
              ],
              "title": "View",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "duplicate_scope",
            "required": false,
            "schema": {
              "default": "canonical",
              "enum": [
                "canonical",
                "aliases",
                "all"
              ],
              "title": "Duplicate Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "canonical_work_item_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Canonical Work Item Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Union_WorkSearchHit__HierarchySummary__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Work"
      },
      "post": {
        "operationId": "create_work_api_v1_projects__project_id__work_items_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkCreation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Work"
      }
    },
...
    "/api/v1/projects/{project_id}/work-items/{work_item_id}": {
      "get": {
        "operationId": "get_work_api_v1_projects__project_id__work_items__work_item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Work"
      },
      "patch": {
        "operationId": "update_work_api_v1_projects__project_id__work_items__work_item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkItemPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Work"
      }
    },
...
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/children": {
      "get": {
        "operationId": "list_children_api_v1_projects__project_id__work_items__work_item_id__children_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "active",
                "dropped",
                "deferred",
                "done",
                "wont-do",
                "promoted",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated",
              "enum": [
                "updated",
                "created",
                "priority"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "source_client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Client"
            }
          },
          {
            "in": "query",
            "name": "source_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Session Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_HierarchySummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Children"
      }
    },
...
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/complete": {
      "post": {
        "operationId": "complete_work_api_v1_projects__project_id__work_items__work_item_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkCompletionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkCompletionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Work"
      }
    },
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/context": {
      "get": {
        "operationId": "recall_work_api_v1_projects__project_id__work_items__work_item_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "recent_limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 0,
              "title": "Recent Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "recent_event_limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 0,
              "title": "Recent Event Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkContext"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recall Work"
      }
    },
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/defer": {
      "post": {
        "description": "Human dashboard action; intentionally absent from the agent MCP surface.",
        "operationId": "defer_work_api_v1_projects__project_id__work_items__work_item_id__defer_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkDeferralCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Defer Work"
      }
    },
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/delete": {
      "post": {
        "operationId": "delete_work_api_v1_projects__project_id__work_items__work_item_id__delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkDeletionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkDeletionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Work"
      }
    },
...
  },
...
}